Moved event and timer_event to events
